Page admin

+++ Controllers/PostsController.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -19,11 +21,15 @@ namespace TP_WS1.Controllers
         }
 
         // GET: Posts
-        public async Task<IActionResult> Index(int? id)
+        public async Task<IActionResult> Index(int? id, DateTime? dateTime)
         {
             if (id != null)
             {
                 var tp1Ws1JeuxVideoContext = _context.Posts.Include(p => p.Game).Include(p => p.User).Where(p => p.GameId == id);
+                return View(await tp1Ws1JeuxVideoContext.ToListAsync());
+            } else if(dateTime != null)
+            {
+                var tp1Ws1JeuxVideoContext = _context.Posts.Include(p => p.Game).Include(p => p.User).Where(p => p.UpdatedAt == dateTime);
                 return View(await tp1Ws1JeuxVideoContext.ToListAsync());
             }
             else
@@ -35,26 +41,11 @@ namespace TP_WS1.Controllers
 
         }
 
-        // GET: Posts/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
 
-            var post = await _context.Posts
-                .Include(p => p.Game)
-                .Include(p => p.User)
-                .FirstOrDefaultAsync(m => m.PostId == id);
-            if (post == null)
-            {
-                return NotFound();
-            }
 
-            return View(post);
-        }
 
+
+        [Authorize]
         // GET: Posts/Create
         public IActionResult Create()
         {
@@ -66,10 +57,12 @@ namespace TP_WS1.Controllers
         // POST: Posts/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("PostId,Title,Message,UserId,GameId,ReactionId,IsArchived")] Post post)
         {
+
             if (ModelState.IsValid)
             {
                 _context.Add(post);
@@ -81,23 +74,42 @@ namespace TP_WS1.Controllers
             return View(post);
         }
 
-        // GET: Posts/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
-            {
                 return NotFound();
-            }
+
 
             var post = await _context.Posts.FindAsync(id);
             if (post == null)
-            {
                 return NotFound();
+
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+  
+            if (currentUserId == post.UserId || User.IsInRole("Admin"))
+            {
+                ViewData["GameId"] = new SelectList(
+                    _context.Games,
+                    "GameId",
+                    "Name",     
+                    post.GameId
+                );
+                ViewData["UserId"] = new SelectList(
+                    _context.AspNetUsers,
+                    "Id",      
+                    "UserName",  
+                    post.UserId
+                );
+                return View(post);
             }
-            ViewData["GameId"] = new SelectList(_context.Games, "GameId", "GameId", post.GameId);
-            ViewData["UserId"] = new SelectList(_context.AspNetUsers, "UserId", "UserId", post.UserId);
-            return View(post);
+            else
+            {
+                // redirige vers l’Index du même contrôleur en passant l’id
+                return RedirectToAction(nameof(Index), new { id });
+            }
         }
+
 
         // POST: Posts/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
@@ -106,54 +118,74 @@ namespace TP_WS1.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("PostId,Title,Message,UserId,GameId,ReactionId,IsArchived")] Post post)
         {
-            if (id != post.PostId)
+            var userIdPost = await _context.Posts.Where(p => p.PostId == id).Select(p => p.UserId).FirstOrDefaultAsync();
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (currentUserId == userIdPost || User.IsInRole("Admin"))
             {
-                return NotFound();
-            }
+                if (id != post.PostId)
+                {
+                    return NotFound();
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(post);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!PostExists(post.PostId))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(post);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!PostExists(post.PostId))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+
+                ViewData["GameId"] = new SelectList(_context.Games, "GameId", "GameId", post.GameId);
+                ViewData["UserId"] = new SelectList(_context.AspNetUsers, "UserId", "UserId", post.UserId);
+                return View(post);
             }
-            ViewData["GameId"] = new SelectList(_context.Games, "GameId", "GameId", post.GameId);
-            ViewData["UserId"] = new SelectList(_context.AspNetUsers, "UserId", "UserId", post.UserId);
-            return View(post);
+            else
+            {
+                return RedirectToAction(nameof(Index), new { id = id });
+            }
         }
 
         // GET: Posts/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            var userIdPost = await _context.Posts.Where(p => p.PostId == id).Select(p => p.UserId).FirstOrDefaultAsync();
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (currentUserId == userIdPost || User.IsInRole("Admin"))
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var post = await _context.Posts
-                .Include(p => p.Game)
-                .Include(p => p.User)
-                .FirstOrDefaultAsync(m => m.PostId == id);
-            if (post == null)
+                var post = await _context.Posts
+                    .Include(p => p.Game)
+                    .Include(p => p.User)
+                    .FirstOrDefaultAsync(m => m.PostId == id);
+                if (post == null)
+                {
+                    return NotFound();
+                }
+
+                return View(post);
+            }
+            else
             {
-                return NotFound();
+                return RedirectToAction(nameof(Index), new { id = id });
             }
-
-            return View(post);
         }
 
         // POST: Posts/Delete/5
@@ -161,14 +193,23 @@ namespace TP_WS1.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var post = await _context.Posts.FindAsync(id);
-            if (post != null)
+            var userIdPost = await _context.Posts.Where(p => p.PostId == id).Select(p => p.UserId).FirstOrDefaultAsync();
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (currentUserId == userIdPost || User.IsInRole("Admin"))
             {
-                _context.Posts.Remove(post);
-            }
+                var post = await _context.Posts.FindAsync(id);
+                if (post != null)
+                {
+                    _context.Posts.Remove(post);
+                }
 
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                return RedirectToAction(nameof(Index), new { id = id });
+            }
         }
 
         private bool PostExists(int id)
